Refactor

+++ src/users/users.controller.ts
@@ -1,11 +1,12 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Query, UseGuards, SerializeOptions, UseInterceptors, ClassSerializerInterceptor } from '@nestjs/common';
-import { ApiOkResponse, ApiQuery, ApiTags, ApiCreatedResponse, ApiBearerAuth } from '@nestjs/swagger';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Query, UseGuards, SerializeOptions, UseInterceptors, ClassSerializerInterceptor, Request, Req, Res, HttpException, HttpStatus } from '@nestjs/common';
+import { ApiQuery, ApiTags, ApiBearerAuth } from '@nestjs/swagger';
+import { JsonApiResponse } from 'src/common/decorators/api.decorators';
 
 // Users
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { FindAllUserResponseDto, UserAttributesEmailDto, UserDto, UserResponseDto } from './dto/find-user.dto';
+import { UserDto, UserEmailDto } from './dto/user.dto';
 
 // Authentication
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
@@ -21,7 +22,7 @@ export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Post()
-  @ApiCreatedResponse({ type: UserResponseDto, description: 'Created user response' })
+  @JsonApiResponse(UserDto)
   async create(@Body() createUserDto: CreateUserDto) {
     return await this.usersService.create(createUserDto);
   }
@@ -32,7 +33,7 @@ export class UsersController {
   @ApiBearerAuth()
   @ApiQuery(decoratorConstant.apiQuery.perPage)
   @ApiQuery(decoratorConstant.apiQuery.page)
-  @ApiOkResponse({ type: FindAllUserResponseDto, description: 'List of all users' })
+  @JsonApiResponse([UserDto])
   async findAll(
     @Query('page') page = decoratorConstant.apiQuery.page.default,
     @Query('perPage') perPage = decoratorConstant.apiQuery.perPage.default,
@@ -41,30 +42,42 @@ export class UsersController {
     return await this.usersService.findAll(pagination);
   }
 
-  @Get(':email')
-  // @Roles('ADMIN')
-  // @UseGuards(JwtAuthGuard, RolesGuard)
-  // @ApiBearerAuth()
-  // @UseInterceptors(ClassSerializerInterceptor)
-  @SerializeOptions({ type: UserDto })
-  @ApiOkResponse({ type: UserResponseDto, description: 'User details' })
-  async findOne(@Param() params: UserAttributesEmailDto) {
-    return await this.usersService.findOne(params.email);
-  }
-
   @Patch(':email')
-  @UseGuards(JwtAuthGuard)
+  @Roles('ADMIN', 'USER')
+  @UseGuards(JwtAuthGuard, RolesGuard)
   @ApiBearerAuth()
-  @ApiOkResponse({ type: UserResponseDto, description: 'User details' })
-  async update(@Param() params: UserAttributesEmailDto, @Body() updateUserDto: UpdateUserDto) {
-    return await this.usersService.update(params.email, updateUserDto);
+  @JsonApiResponse(UserDto)
+  async update(@Req() request, @Body() updateUserDto: UpdateUserDto, @Param() params: UserEmailDto) {
+    if (request.user.role === 'ADMIN') {
+      return await this.usersService.update(params.email, updateUserDto);
+    }
+    if (params.email === request.user.emai) {
+      return await this.usersService.update(request.user.email, updateUserDto);
+    }
+
+    throw new HttpException('Please insert your current email in params section.', HttpStatus.FORBIDDEN);
   }
 
   @Delete(':email')
-  @UseGuards(JwtAuthGuard)
+  @Roles('ADMIN')
+  @UseGuards(JwtAuthGuard, RolesGuard)
   @ApiBearerAuth()
-  @ApiOkResponse({ type: UserResponseDto, description: 'User removed' })
-  async remove(@Param() params: UserAttributesEmailDto) {
+  @JsonApiResponse(UserDto)
+  async remove(@Req() request, @Param() params: UserEmailDto) {
+    if (request.user.email === params.email) {
+      throw new HttpException('You can not delete yourself!', HttpStatus.FORBIDDEN);
+    }
     return await this.usersService.remove(params.email);
   }
+
+  @Get('detail')
+  @Roles('ADMIN', 'USER')
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  @ApiBearerAuth()
+  @JsonApiResponse(UserDto)
+  // TODO add type for request
+  async findOne(@Req() request) {
+    return await this.usersService.findOne(request.user.id);
+  }
+
 }
